Extract pointer handling from AppElement into an internal controller
AppElement carried a self-contained pointer-input subsystem - picker
lifecycle, canvas handlers, listener synchronization, coordinate
mapping, camera selection, hover tracking, ordered dispatch and click
synthesis - alongside application boot, sizing, preload and teardown.

The subsystem moves verbatim to an internal PointerController with a
small explicit lifecycle: connect once the application and canvas
exist, resize with the drawing buffer, syncListeners on descendant
listener changes, and disconnect on teardown. The controller reads two
host services instead of AppElement internals: resolving a graph node
to its element, and the current pointer-target elements. AppElement
keeps boot/teardown, sizing and the element-registration contract,
delegating pointer work through four thin calls.

Behavior is unchanged: the existing pointer integration suite passes
as-is apart from repointing its two private-internals seams at the
controller, and focused controller tests cover what the extraction
makes testable directly - the lifecycle around connect, and the
demand-driven canvas listener table exercised through stub hosts
without any entity elements.

The controller is internal: excluded from the Custom Elements Manifest
like the other non-element modules and stripped from the emitted
declarations, which stay byte-identical along with the manifest and
editor integrations.

Fixes #440
